+++ 1682/gas_male/gasmaleperf.py
@@ -1,8 +1,9 @@
 """Jungle Hawk Owl"""
 import numpy as np
+from submodels.breguet_endurance import BreguetEndurance
+from submodels.flight_state import FlightState
+from submodels.gas_engine import Engine
 from gpkit import Model, Variable, vectorize, units
-from gpkit.tools import te_exp_minus1
-from gpkit.constraints.tight import TightConstraintSet as TCS
 
 # pylint: disable=invalid-name
 
@@ -104,51 +105,6 @@ class AircraftPerf(Model):
 
         Model.__init__(self, None, [self.dynamicmodels, constraints], **kwargs)
 
-class FlightState(Model):
-    "One chunk of a mission"
-    def __init__(self, alt, onStation, wind, **kwargs):
-
-        self.onStation = onStation
-
-        mu = Variable("\\mu", 1.628e-5, "N*s/m^2", "dynamic viscosity")
-        rho = Variable("\\rho", "kg/m^3", "air density")
-        h = Variable("h", alt, "ft", "altitude")
-        href = Variable("h_{ref}", 15000, "ft", "Reference altitude")
-        psl = Variable("p_{sl}", 101325, "Pa", "Pressure at sea level")
-        Latm = Variable("L_{atm}", 0.0065, "K/m", "Temperature lapse rate")
-        Tsl = Variable("T_{sl}", 288.15, "K", "Temperature at sea level")
-        temp = [(t.value - l.value*v.value).magnitude
-                for t, v, l in zip(Tsl, h, Latm)]
-        Tatm = Variable("t_{atm}", temp, "K", "Air temperature")
-        mu = Variable("\\mu", "N*s/m^2", "Dynamic viscosity")
-        musl = Variable("\\mu_{sl}", 1.789*10**-5, "N*s/m^2",
-                        "Dynamic viscosity at sea level")
-        Rspec = Variable("R_{spec}", 287.058, "J/kg/K",
-                         "Specific gas constant of air")
-
-        # Atmospheric variation with altitude (valid from 0-7km of altitude)
-        constraints = [rho == psl*Tatm**(5.257-1)/Rspec/(Tsl**5.257),
-                       (mu/musl)**0.1 == 0.991*(h/href)**(-0.00529),
-                       h == h,
-                       href == href]
-
-        V = Variable("V", "m/s", "true airspeed")
-
-        if wind:
-
-            V_wind = Variable("V_{wind}", 25, "m/s", "Wind speed")
-            constraints.extend([V >= V_wind])
-
-        else:
-
-            V_wind = Variable("V_{wind}", "m/s", "Wind speed")
-            V_ref = Variable("V_{ref}", 25, "m/s", "Reference wind speed")
-
-            constraints.extend([(V_wind/V_ref) >= 0.6462*(h/href) + 0.3538,
-                                V >= V_wind])
-        Model.__init__(self, None, constraints, **kwargs)
-
-
 class FlightSegment(Model):
     "creates flight segment for aircraft"
     def __init__(self, N, aircraft, alt=15000, onStation=False, wind=False,
@@ -222,24 +178,6 @@ class Climb(Model):
 
         Model.__init__(self, None, [fs, constraints], **kwargs)
 
-class BreguetEndurance(Model):
-    "breguet endurance model"
-    def __init__(self, perf, **kwargs):
-        z_bre = Variable("z_{bre}", "-", "Breguet coefficient")
-        t = Variable("t", "days", "Time per flight segment")
-        f_fueloil = Variable("f_{(fuel/oil)}", 0.98, "-", "Fuel-oil fraction")
-        Wfuel = Variable("W_{fuel}", "lbf", "Segment-fuel weight")
-        g = Variable("g", 9.81, "m/s^2", "gravitational acceleration")
-
-        constraints = [
-            TCS([z_bre >= (perf["P_{total}"]*t*perf["BSFC"]*g
-                           / (perf["W_{end}"]*perf["W_{start}"])**0.5)]),
-            f_fueloil*Wfuel/perf["W_{end}"] >= te_exp_minus1(z_bre, 3),
-            perf["W_{start}"] >= perf["W_{end}"] + Wfuel
-            ]
-
-        Model.__init__(self, None, constraints, **kwargs)
-
 class SteadyLevelFlight(Model):
     "steady level flight model"
     def __init__(self, state, aircraft, perf, etap, **kwargs):
@@ -254,92 +192,6 @@ class SteadyLevelFlight(Model):
             T >= (0.5*state["\\rho"]*state["V"]**2*perf["C_D"]
                   *aircraft.wing["S"]),
             perf["P_{shaft}"] >= T*state["V"]/etaprop]
-
-        Model.__init__(self, None, constraints, **kwargs)
-
-class Engine(Model):
-    "engine model"
-    def __init__(self, DF70=False, **kwargs):
-
-        self.DF70 = DF70
-
-        W = Variable("W", "lbf", "Installed/Total engine weight")
-        mfac = Variable("m_{fac}", 1.0, "-", "Engine weight margin factor")
-
-        if DF70:
-            Wdf70 = Variable("W_{DF70}", 7.1, "lbf",
-                             "Installed/Total DF70 engine weight")
-            Pslmax = Variable("P_{sl-max}", 5.17, "hp",
-                              "Max shaft power at sea level")
-            constraints = [W/mfac >= Wdf70,
-                           Pslmax == Pslmax]
-
-        else:
-            Pref = Variable("P_{ref}", 2.295, "hp", "Reference shaft power")
-            Wengref = Variable("W_{eng-ref}", 4.4107, "lbf",
-                               "Reference engine weight")
-            Weng = Variable("W_{eng}", "lbf", "engine weight")
-            Pslmax = Variable("P_{sl-max}", "hp",
-                              "Max shaft power at sea level")
-
-            constraints = [
-                Weng/Wengref >= 0.5538*(Pslmax/Pref)**1.075,
-                W/mfac >= 2.572*Weng**0.922*units("lbf")**0.078]
-
-        self.flight_model = EnginePerf
-
-        Model.__init__(self, None, constraints, **kwargs)
-
-class EnginePerf(Model):
-    "engine performance model"
-    def __init__(self, static, state, **kwargs):
-
-        Pshaft = Variable("P_{shaft}", "hp", "Shaft power")
-        bsfc = Variable("BSFC", "lb/hr/hp", "Brake specific fuel consumption")
-        rpm = Variable("RPM", "rpm", "Engine operating RPM")
-        Pavn = Variable("P_{avn}", 40, "watts", "Avionics power")
-        Ppay = Variable("P_{pay}", 10, "watts", "Payload power")
-        Ptotal = Variable("P_{total}", "hp", "Total power, avionics included")
-        eta_alternator = Variable("\\eta_{alternator}", 0.8, "-",
-                                  "alternator efficiency")
-        lfac = [1 - 0.906**(1/0.15)*(v.value/hr.value)**0.92
-                for v, hr in zip(state["h"], state["h_{ref}"])]
-        Leng = Variable("L_{eng}", lfac, "-", "shaft power loss factor")
-        Pshaftmax = Variable("P_{shaft-max}",
-                             "hp", "Max shaft power at altitude")
-        mfac = Variable("m_{fac}", 1.0, "-", "BSFC margin factor")
-
-        if static.DF70:
-            rpm_max = Variable("RPM_{max}", 7698, "rpm", "Maximum RPM")
-            bsfc_min = Variable("BSFC_{min}", 0.3162, "kg/kW/hr",
-                                "Minimum BSFC")
-
-            constraints = [
-                (bsfc/mfac/bsfc_min)**35.7 >=
-                (2.29*(rpm/rpm_max)**8.02 + 0.00114*(rpm/rpm_max)**-38.3),
-                (Ptotal/Pshaftmax)**0.1 == 0.999*(rpm/rpm_max)**0.294
-                ]
-        else:
-            bsfc_min = Variable("BSFC_{min}", 0.32, "kg/kW/hr",
-                                "Minimum BSFC")
-            rpm_max = Variable("RPM_{max}", 9000, "rpm", "Maximum RPM")
-
-            constraints = [
-                (bsfc/mfac/bsfc_min)**0.129 >=
-                (0.972*(rpm/rpm_max)**-0.141 + 0.0268*(rpm/rpm_max)**9.62),
-                (Ptotal/Pshaftmax)**0.1 == 0.999*(rpm/rpm_max)**0.292
-                ]
-
-        constraints.extend([Pshaftmax/static["P_{sl-max}"] == Leng,
-                            Pshaftmax >= Ptotal,
-                            rpm <= rpm_max])
-
-        if state.onStation:
-            constraints.extend([
-                Ptotal >= Pshaft + (Pavn + Ppay)/eta_alternator])
-        else:
-            constraints.extend([Ptotal >= Pshaft + Pavn/eta_alternator])
-
 
         Model.__init__(self, None, constraints, **kwargs)
