+++ 1682/gas_male/submodels/__init__.py
@@ -1,0 +1,1 @@
+from breguet_endurance import BreguetEndurance
